+++ src/fast-check-default.ts
@@ -46,7 +46,9 @@ import { dictionary } from './arbitrary/dictionary';
 import { emailAddress } from './check/arbitrary/EmailArbitrary';
 import { double, float, DoubleConstraints, FloatConstraints } from './check/arbitrary/FloatingPointArbitrary';
 import { frequency, WeightedArbitrary, FrequencyValue, FrequencyContraints } from './arbitrary/frequency';
-import { compareBooleanFunc, compareFunc, func } from './check/arbitrary/FunctionArbitrary';
+import { compareBooleanFunc } from './arbitrary/compareBooleanFunc';
+import { compareFunc } from './arbitrary/compareFunc';
+import { func } from './arbitrary/func';
 import { domain } from './check/arbitrary/HostArbitrary';
 import { integer, IntegerConstraints } from './arbitrary/integer';
 import { maxSafeInteger } from './arbitrary/maxSafeInteger';
